doc['ul'].value syntax
