+++ server/src/main/java/org/elasticsearch/script/Converters.java
@@ -116,6 +116,9 @@ public class Converters {
         };
     }
 
+    // No instances, please
+    private Converters() {}
+
     public static BigIntegerField LongToBigInteger(LongField sourceField) {
         FieldValues<Long> fv = sourceField.getFieldValues();
         return new BigIntegerField(sourceField.getName(), new DelegatingFieldValues<>(fv) {
@@ -288,5 +291,37 @@ public class Converters {
                 }
             }
         });
+    }
+
+    /**
+     * Helper for creating {@link Converter} classes which delegates all un-overridden methods to the underlying
+     * {@link FieldValues}.
+     */
+    public abstract static class DelegatingFieldValues<T, D> implements FieldValues<T> {
+        protected FieldValues<D> values;
+
+        public DelegatingFieldValues(FieldValues<D> values) {
+            this.values = values;
+        }
+
+        @Override
+        public boolean isEmpty() {
+            return values.isEmpty();
+        }
+
+        @Override
+        public int size() {
+            return values.size();
+        }
+
+        @Override
+        public long getLongValue() {
+            return values.getLongValue();
+        }
+
+        @Override
+        public double getDoubleValue() {
+            return values.getDoubleValue();
+        }
     }
 }
