+++ x-pack/plugin/mapper-unsigned-long/src/main/java/org/elasticsearch/xpack/unsignedlong/DocValuesWhitelistExtension.java
@@ -9,17 +9,43 @@ package org.elasticsearch.xpack.unsignedlong;
 
 import org.elasticsearch.painless.spi.PainlessExtension;
 import org.elasticsearch.painless.spi.Whitelist;
+import org.elasticsearch.painless.spi.WhitelistLoader;
+import org.elasticsearch.script.AggregationScript;
+import org.elasticsearch.script.BucketAggregationSelectorScript;
+import org.elasticsearch.script.FieldScript;
+import org.elasticsearch.script.FilterScript;
+import org.elasticsearch.script.NumberSortScript;
+import org.elasticsearch.script.ScoreScript;
 import org.elasticsearch.script.ScriptContext;
+import org.elasticsearch.script.StringSortScript;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+import static java.util.Collections.singletonList;
+
 public class DocValuesWhitelistExtension implements PainlessExtension {
+
+    private static final Whitelist WHITELIST = WhitelistLoader.loadFromResourceFiles(DocValuesWhitelistExtension.class, "whitelist.txt");
 
     @Override
     public Map<ScriptContext<?>, List<Whitelist>> getContextWhitelists() {
-        // TODO: support unsigned_long in scripts
-        return Collections.emptyMap();
+        List<Whitelist> whitelist = singletonList(WHITELIST);
+        return Map.of(
+            FieldScript.CONTEXT,
+            whitelist,
+            ScoreScript.CONTEXT,
+            whitelist,
+            FilterScript.CONTEXT,
+            whitelist,
+            AggregationScript.CONTEXT,
+            whitelist,
+            NumberSortScript.CONTEXT,
+            whitelist,
+            StringSortScript.CONTEXT,
+            whitelist,
+            BucketAggregationSelectorScript.CONTEXT,
+            whitelist
+        );
     }
 }
