+++ x-pack/plugin/src/yamlRestTest/resources/rest-api-spec/test/unsigned_long/50_script_values.yml
@@ -46,6 +46,36 @@ setup:
   - match: { hits.hits.2.fields.scripted_ul.0: -9223372036854775808 }
   - match: { hits.hits.3.fields.scripted_ul.0: 9223372036854775807 }
   - match: { hits.hits.4.fields.scripted_ul.0: 0 }
+  - do:
+      search:
+        index: test1
+        body:
+          sort: [ { ul: desc } ]
+          script_fields:
+            scripted_ul:
+              script:
+                source: "field('ul').getLong(1000L)"
+
+  - match: { hits.hits.0.fields.scripted_ul.0: -1 }
+  - match: { hits.hits.1.fields.scripted_ul.0: -3 }
+  - match: { hits.hits.2.fields.scripted_ul.0: -9223372036854775808 }
+  - match: { hits.hits.3.fields.scripted_ul.0: 9223372036854775807 }
+  - match: { hits.hits.4.fields.scripted_ul.0: 0 }
+  - do:
+      search:
+        index: test1
+        body:
+          sort: [ { ul: desc } ]
+          script_fields:
+            scripted_ul:
+              script:
+                source: "doc['ul'].value"
+
+  - match: { hits.hits.0.fields.scripted_ul.0: -1 }
+  - match: { hits.hits.1.fields.scripted_ul.0: -3 }
+  - match: { hits.hits.2.fields.scripted_ul.0: -9223372036854775808 }
+  - match: { hits.hits.3.fields.scripted_ul.0: 9223372036854775807 }
+  - match: { hits.hits.4.fields.scripted_ul.0: 0 }
 
 ---
 "Scripted sort values":
@@ -65,6 +95,41 @@ setup:
   - match: { hits.hits.2.sort: [9.223372036854776E18] }
   - match: { hits.hits.3.sort: [9.223372036854776E18] }
   - match: { hits.hits.4.sort: [0.0] }
+  - do:
+      search:
+        index: test1
+        body:
+          sort:
+            _script:
+              order: desc
+              type: number
+              script:
+                source: "doc['ul'].value"
+
+
+  - match: { hits.hits.0.sort: [9.223372036854776E18] }
+  - match: { hits.hits.1.sort: [0.0] }
+  - match: { hits.hits.2.sort: [-1.0] }
+  - match: { hits.hits.3.sort: [-3.0] }
+  - match: { hits.hits.4.sort: [-9.223372036854776E18] }
+---
+"Scripted sort values via doc":
+  - do:
+      search:
+        index: test1
+        body:
+          sort:
+            _script:
+              order: desc
+              type: number
+              script:
+                source: "doc['ul'].value"
+
+  - match: { hits.hits.0.sort: [9.223372036854776E18] }
+  - match: { hits.hits.1.sort: [0.0] }
+  - match: { hits.hits.2.sort: [-1.0] }
+  - match: { hits.hits.3.sort: [-3.0] }
+  - match: { hits.hits.4.sort: [-9.223372036854776E18] }
 
 ---
 "Script query":
@@ -95,6 +160,31 @@ setup:
                   script:
                     source: "field('ul').isEmpty() == false"
   - match: { hits.total.value: 5 }
+  - do:
+      search:
+        index: test1
+        body:
+          query:
+            bool:
+              filter:
+                script:
+                  script:
+                    source: "doc['ul'].value.doubleValue() > 9E18"
+  - match: { hits.total.value: 1 }
+  - match: { hits.hits.0._id: "2" }
+
+  - do:
+      search:
+        index: test1
+        body:
+          size: 0
+          query:
+            bool:
+              filter:
+                script:
+                  script:
+                    source: "doc['ul'].size() > 0"
+  - match: { hits.total.value: 5 }
 
 ---
 "script_score query":
@@ -107,5 +197,16 @@ setup:
               query: {match_all: {}}
               script:
                 source: "field('ul').as(Field.BigInteger).getValue(BigInteger.valueOf(Long.parseUnsignedLong('18446744073709551614'))).doubleValue()"
+
+  - match: { hits.total.value: 5 }
+  - do:
+      search:
+        index: test1
+        body:
+          query:
+            script_score:
+              query: {match_all: {}}
+              script:
+                source: "Math.abs(doc['ul'].value)"
 
   - match: { hits.total.value: 5 }
